requirements.txt
# --------- pytorch --------- #
torch>=1.8.1
torchvision>=0.9.1
pytorch-lightning>=1.3.8


# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0
torchmetrics==0.5.1
torch-fidelity==0.3.0
# hydra-ax-sweeper
# hydra-ray-launcher
# hydra-submitit-launcher

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis

# --------- others --------- #
python-dotenv   # loading env variables from .env file
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
scikit-learn    # used in some callbacks
seaborn         # used in some callbacks
jupyterlab      # better jupyter notebooks
pudb            # debugger
